move the journey display into a separate file
